perf: add confirm when paste、delete、clear records

+++ apps/nextjs-app/src/features/app/blocks/view/grid/GridViewBaseInner.tsx
@@ -78,7 +78,7 @@ import {
   useRecordOperations,
   useButtonClickStatus,
 } from '@teable/sdk/hooks';
-import { ConfirmDialog, useToast } from '@teable/ui-lib';
+import { ConfirmDialog, useConfirm, useToast } from '@teable/ui-lib';
 import { toast as sonnerToast } from '@teable/ui-lib/shadcn/ui/sonner';
 import { isEqual, keyBy, uniqueId, groupBy } from 'lodash';
 import { useRouter } from 'next/router';
@@ -392,6 +392,8 @@ export const GridViewBaseInner: React.FC<IGridViewBaseInnerProps> = (
     [recordMap, columns, t]
   );
 
+  const { confirm } = useConfirm();
+
   // eslint-disable-next-line sonarjs/cognitive-complexity
   const onContextMenu = (selection: CombinedSelection, position: IPosition) => {
     const { isCellSelection, isRowSelection, isColumnSelection, ranges } = selection;
@@ -417,6 +419,24 @@ export const GridViewBaseInner: React.FC<IGridViewBaseInnerProps> = (
           position,
           isMultipleSelected,
           deleteRecords: async () => {
+            const [startRange, endRange] = selection.ranges;
+            const [, startRow] = startRange;
+            const [, endRow] = endRange;
+            const deleteRows = endRow - startRow + 1;
+
+            if (deleteRows >= 10) {
+              const confirmed = await confirm({
+                title: t('table:table.actionTips.deleteRecordConfirmTitle'),
+                description: t('table:table.actionTips.deleteRecordConfirmDescription', {
+                  recordCount: deleteRows,
+                }),
+                confirmText: t('table:table.actionTips.deleteRecord'),
+                cancelText: t('common:actions.cancel'),
+                confirmButtonVariant: 'destructive',
+              });
+              if (!confirmed) return;
+            }
+
             deleteRecords(selection);
             gridRef.current?.setSelection(emptySelection);
           },
